requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fastapi==0.104.1
googleapis-common-protos==1.61.0
grpcio==1.59.3
h11==0.14.0
httptools==0.6.1
idna==3.4
importlib-metadata==6.8.0
markdown-it-py==3.0.0
mdurl==0.1.2
opentelemetry-api==1.21.0
opentelemetry-distro==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-exporter-richconsole==0.42b0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-aws-lambda==0.42b0
opentelemetry-instrumentation-dbapi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-grpc==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlite3==0.42b0
opentelemetry-instrumentation-tortoiseorm==0.42b0
opentelemetry-instrumentation-urllib==0.42b0
opentelemetry-instrumentation-urllib3==0.42b0
opentelemetry-instrumentation-wsgi==0.42b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
protobuf==4.25.1
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
zipp==3.17.0
